Increasing the grid size
I have converted my single vector of cells into a deque of deques of cells which allows me to add more cells to both the front and back of the existing elements.

I have also created a set of functions which add cells to the deques in all directions (up, down, left and right); although, this is working effectively I have bound their spawining to the space button.

Next, I plan to store the upper/lower bounds of the grid to allow for dynamic spawning when the player gets within range of those bounds.