Basic Attack
I have implemented a basic attack which can be held endlessly, for now this is fine  but will need a limit later.

Next I will need to add enemies into the game, I will start with stationary ones which will spawn randomly within the game world so I can work on the random genetration before worig with their interaction with the player's attack.