Attacking, knocking back and killing enemies
I have implemented a full attacking system into the game which allows for the player to attack and kill an enemy in the game world. If the player hits the enemy with an attack (using the left mouse click) then the enemy will be knocked back in the direction of the attack and lose some life; if the enemy loses all of their life then they will be killed and removed from the game. #

Next I will need to implement a way for the enemies to attack the player, a simple way would be for the enemy to collide with the player rather than attacking in a similar way to the player.

Some experimentation is needed.