Added Enemy Attacking
I have added a system which allows for the enemies to attack the player if they get within a short distance to the player. There is no attacking animation for them and no lose screen for the game but there is the basis for those.

Next I will try to fix the map, add enemies which randomly spawn into the map and sort out the map generation (maybe start fresh with it).