Expanding map
I have added an expanding map to the game, it is currently a manual expansion using the arrow keys and there is no terrain generation yet.

Next, I will add a bounds to the map which will grow with the map, using this I will be able to detect when the player has reached one of the edges and the map should then expand automatically.

Once that is finished I will need to add in a tile system to be generated.
TBC.