Added map generation
I have enabled the map to generate new tiles on the cells, this is currently limited to only grass and mountain tiles but for now it works fine (after a little while the fps drops, optimization will be needed later but fine for now).

Next, I plan to prevent movement through mountain tiles for both the player and enemies.