Fixed Crashing Issue
I have fixed the issue where in the deque used in the map would get out of range aborting the game. The issue was that I wasn't joining the update thread into the main loop before leaving to the lose screen.

This is fixed. The next step is to comment properly and push it to a porfolio. Posibly update the readme on github as well.